requirements.txt
# Core API
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.1
python-multipart>=0.0.6

# AI & Vector DB
pymilvus>=2.6.0
chromadb>=0.4.15
pageindex==0.2.5.1
# sentence-transformers>=2.2.2  <-- OPTIONAL/LOCAL ONLY
# torch>=2.0.0  <-- Let dependencies resolve or use CPU version implicitly
rank-bm25>=0.2.2

# LLM Clients
openai>=1.3.0
groq>=0.4.0
cerebras-cloud-sdk>=1.0.0
httpx>=0.24.1

# Utilities
numpy<2.0.0
pandas>=1.5.0
tqdm>=4.65.0
regex>=2023.10.3
duckduckgo-search>=6.0.0
requests>=2.31.0
pydantic>=2.0.0
psutil>=5.9.0
# camel-ai>=0.1.0  <-- HEAVY OPTIONAL: install manually if OwlJudicialWorkforce needed
fire>=0.5.0
python-json-logger>=2.0.7
aiofiles>=23.2.1
gunicorn>=21.2.0
